Added libnx-rs bindings behind aarch64 cfg

+++ src/libstd/sys/horizon/net.rs
@@ -344,7 +344,7 @@ impl Socket {
 
     pub fn set_nonblocking(&self, nonblocking: bool) -> io::Result<()> {
         let mut nonblocking = nonblocking as libc::c_int;
-        cvt(unsafe { libc::ioctl(*self.as_inner(), libc::FIONBIO as u32, &mut nonblocking) }).map(|_| ())
+        cvt(unsafe { libc::ioctl(*self.as_inner(), libc::FIONBIO as u64, &mut nonblocking) }).map(|_| ())
     }
 
     pub fn take_error(&self) -> io::Result<Option<io::Error>> {
